requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
arxiv==2.4.0
azure-core==1.38.0
azure-storage-blob==12.28.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
distro==1.9.0
fastapi==0.128.0
feedparser==6.0.12
filelock==3.20.3
fsspec==2026.1.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.1
idna==3.11
isodate==0.7.2
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.6
langchain-core==1.2.7
langgraph==1.0.6
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.6
langgraph-sdk==0.3.3
langsmith==0.6.4
openai==2.15.0
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pypdf==6.6.0
PyPDF2==3.0.1
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
sgmllib3k==1.0.0
shellingham==1.5.4
sniffio==1.3.1
starlette==0.50.0
tenacity==9.1.2
tqdm==4.67.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.13.0
uvicorn==0.40.0
xxhash==3.6.0
zstandard==0.25.0
